Create x/metrics package
