+++ tools/go.mod
@@ -4,13 +4,17 @@ go 1.13
 
 require (
 	github.com/fatih/color v1.9.0
-	github.com/go-redis/redis/v8 v8.11.2
-	github.com/google/uuid v1.2.0
+	github.com/go-redis/redis/v8 v8.11.4
+	github.com/google/uuid v1.3.0
 	github.com/hibiken/asynq v0.17.1
+	github.com/hibiken/asynq/x v0.0.0-00010101000000-000000000000 // indirect
 	github.com/mitchellh/go-homedir v1.1.0
-	github.com/prometheus/client_golang v0.9.3
+	github.com/prometheus/client_golang v1.11.0
 	github.com/spf13/cobra v1.1.1
 	github.com/spf13/viper v1.7.0
 )
 
-replace github.com/hibiken/asynq => ./..
+replace (
+	github.com/hibiken/asynq => ./..
+	github.com/hibiken/asynq/x => ./../x
+)
